Button navigation issue, Needs to be fixed;Comments for next iteration of frontend in files; Moved Navbar to its own folder
